% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{Account}
\alias{Account}
\title{Ad Exchange Seller API Objects 
Accesses the inventory of Ad Exchange seller users and generates reports.}
\usage{
Account(id = NULL, name = NULL)
}
\arguments{
\item{id}{Unique identifier of this account}

\item{name}{Name of this account}
}
\value{
Account object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:31:42
filename: /Users/mark/dev/R/autoGoogleAPI/googleadexchangesellerv11.auto/R/adexchangeseller_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Account Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

