% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{Alerts}
\alias{Alerts}
\title{Alerts Object}
\usage{
Alerts(items = NULL)
}
\arguments{
\item{items}{The alerts returned in this list response}
}
\value{
Alerts object
}
\description{
Alerts Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

