% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adexchangeseller_objects.R
\name{PreferredDeals}
\alias{PreferredDeals}
\title{PreferredDeals Object}
\usage{
PreferredDeals(items = NULL)
}
\arguments{
\item{items}{The preferred deals returned in this list response}
}
\value{
PreferredDeals object
}
\description{
PreferredDeals Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
